// src/models/user.ts
export interface User {
  id: number;
  nombre: string;
  apellidoMaterno: string;
  apellidoPaterno: string;
  cargo: string;
  empresa: string;
  calle: string;
  numeroExt: string;
  numeroInt: string;
  colonia: string;
  municipio: string;
  estado: string;
  codigoPostal: string;
  telefono: string;
  correoElectronico: string;
  fechaNacimiento: Date;
}